fhir-api/requirements/base.txt
elasticsearch==7.0.5
fhirstore @ git+https://github.com/arkhn/pyfhirstore@0.5.0#egg=fhirstore
flask-cors==3.0.8
flask==1.1.1
jsonschema==3.0.2
PyJWT[crypto]==1.7.1
pymongo==3.9.0
pysin==1.5.2
urllib3==1.25.7
uWSGI==2.0.18
uwsgidecorators==1.1.0
pandas==1.0.3
fhir2dataset==0.1.2
arkhn_arx==0.0.8
python-dotenv==0.14.0